Oracle: Formatting

+++ src/Oracle.sol
@@ -26,8 +26,8 @@ import {IOracle} from "./interfaces/IOracle.sol";
  */
 contract Oracle is TSOwnable, IOracle {
     struct Report {
-        uint256 timestamp;
-        uint256 payload;
+        uint timestamp;
+        uint payload;
     }
 
     //--------------------------------------------------------------------------
@@ -48,7 +48,7 @@ contract Oracle is TSOwnable, IOracle {
     /// @param payload The payload of the report.
     /// @param timestamp The timestamp of the report.
     event ProviderReportPushed(
-        address indexed provider, uint256 payload, uint256 timestamp
+        address indexed provider, uint payload, uint timestamp
     );
 
     /// @notice Emitted when reports from a provider are purged.
@@ -70,7 +70,7 @@ contract Oracle is TSOwnable, IOracle {
     /// @param oldMinimumProviders The old number of minimum providers.
     /// @param newMinimumProviders The new number of minimum providers.
     event MinimumProvidersChanged(
-        uint256 oldMinimumProviders, uint256 newMinimumProviders
+        uint oldMinimumProviders, uint newMinimumProviders
     );
 
     /// @notice Emitted when the oracle is marked as invalid.
@@ -83,12 +83,12 @@ contract Oracle is TSOwnable, IOracle {
     // Storage
 
     /// @notice The number of seconds after which a report is deemed expired.
-    uint256 public immutable reportExpirationTime;
+    uint public immutable reportExpirationTime;
 
     /// @notice The number of seconds since reporting that has to pass before
     ///         a report is usable.
     /// @dev Gives a buffer to purge reports in case of faulty payload.
-    uint256 public immutable reportDelay;
+    uint public immutable reportDelay;
 
     /// @notice Addresses of providers authorized to push reports.
     /// @dev Changeable by owner.
@@ -101,7 +101,7 @@ contract Oracle is TSOwnable, IOracle {
     /// @notice The minimum number of providers with valid reports to consider
     ///         the aggregate report valid.
     /// @dev Changeable by owner.
-    uint256 public minimumProviders;
+    uint public minimumProviders;
 
     /// @notice Flag to indicate if oracle delivers correct value.
     /// @dev Changeable by owner.
@@ -111,9 +111,9 @@ contract Oracle is TSOwnable, IOracle {
     // Constructor
 
     constructor(
-        uint256 reportExpirationTime_,
-        uint256 reportDelay_,
-        uint256 minimumProviders_
+        uint reportExpirationTime_,
+        uint reportDelay_,
+        uint minimumProviders_
     ) {
         // Make sure that two reports from same provider can be valid.
         require(reportExpirationTime_ > 2 * reportDelay_);
@@ -138,10 +138,9 @@ contract Oracle is TSOwnable, IOracle {
     /// @dev Only callable if at least reportDelay seconds passed since last
     ///      report.
     /// @param payload The report's data.
-    function pushReport(uint256 payload) external {
+    function pushReport(uint payload) external {
         Report[2] storage reports = providerReports[msg.sender];
-        uint256[2] memory timestamps =
-            [reports[0].timestamp, reports[1].timestamp];
+        uint[2] memory timestamps = [reports[0].timestamp, reports[1].timestamp];
 
         // Check if provider exists.
         if (timestamps[0] == 0) {
@@ -184,24 +183,19 @@ contract Oracle is TSOwnable, IOracle {
 
     /// @notice Returns the oracle data and a boolean indicating if data is
     ///         valid.
-    function getData()
-        external
-        view
-        override (IOracle)
-        returns (uint256, bool)
-    {
+    function getData() external view override (IOracle) returns (uint, bool) {
         // Return early if oracle is marked as invalid.
         if (!isValid) {
             return (0, false);
         }
 
-        uint256 reportsCount = providers.length;
-        uint256[] memory validReports = new uint[](reportsCount);
-        uint256 size;
+        uint reportsCount = providers.length;
+        uint[] memory validReports = new uint[](reportsCount);
+        uint size;
 
         // Get for each provider the average of their report payloads.
-        for (uint256 i; i < reportsCount;) {
-            uint256 result;
+        for (uint i; i < reportsCount; ) {
+            uint result;
             bool valid;
             (result, valid) = _getAverageReportPayload(providers[i]);
 
@@ -209,9 +203,7 @@ contract Oracle is TSOwnable, IOracle {
                 validReports[size++] = result;
             }
 
-            unchecked {
-                ++i;
-            }
+            unchecked { ++i; }
         }
 
         // Check that enough reports are used.
@@ -228,7 +220,7 @@ contract Oracle is TSOwnable, IOracle {
 
     /// @notice Returns the number of providers eligible to push reports.
     /// @return uint The number of providers.
-    function providersSize() external view returns (uint256) {
+    function providersSize() external view returns (uint) {
         return providers.length;
     }
 
@@ -257,10 +249,7 @@ contract Oracle is TSOwnable, IOracle {
     ///         deliver reports for the oracle data to be valid.
     /// @dev Only callable by owner.
     /// @param minimumProviders_ The minimum number of providers.
-    function setMinimumProviders(uint256 minimumProviders_)
-        external
-        onlyOwner
-    {
+    function setMinimumProviders(uint minimumProviders_) external onlyOwner {
         // Make sure that at least one provider has to deliver reports.
         require(minimumProviders_ != 0);
 
@@ -315,8 +304,8 @@ contract Oracle is TSOwnable, IOracle {
         delete providerReports[provider];
 
         // Remove provider.
-        uint256 len = providers.length;
-        for (uint256 i; i < len;) {
+        uint len = providers.length;
+        for (uint i; i < len; ) {
             if (providers[i] == provider) {
                 // If not last elem in array, copy last elem to this index.
                 if (i + 1 != len) {
@@ -327,9 +316,7 @@ contract Oracle is TSOwnable, IOracle {
                 break;
             }
 
-            unchecked {
-                ++i;
-            }
+            unchecked { ++i; }
         }
     }
 
@@ -341,23 +328,24 @@ contract Oracle is TSOwnable, IOracle {
     function _getAverageReportPayload(address provider)
         private
         view
-        returns (uint256, bool)
+        returns (uint, bool)
     {
-        uint256 minValidTimestamp = block.timestamp - reportExpirationTime;
-        uint256 maxValidTimestamp = block.timestamp - reportDelay;
+        uint minValidTimestamp = block.timestamp - reportExpirationTime;
+        uint maxValidTimestamp = block.timestamp - reportDelay;
 
         Report[2] memory reports = providerReports[provider];
 
         // Get index of recent and past reports.
-        uint8 indexRecent = reports[0].timestamp >= reports[1].timestamp ? 0 : 1;
+        uint8 indexRecent = reports[0].timestamp >= reports[1].timestamp
+                            ? 0
+                            : 1;
         uint8 indexPast = 1 - indexRecent;
 
         // Cache timestamp and payload from reports.
-        uint256 recentReportTimestamp = reports[indexRecent].timestamp;
-        uint256 pastReportTimestamp = reports[indexPast].timestamp;
-        uint256 recentReportPayload =
-            providerReports[provider][indexRecent].payload;
-        uint256 pastReportPayload = providerReports[provider][indexPast].payload;
+        uint recentReportTimestamp = reports[indexRecent].timestamp;
+        uint pastReportTimestamp = reports[indexPast].timestamp;
+        uint recentReportPayload = providerReports[provider][indexRecent].payload;
+        uint pastReportPayload = providerReports[provider][indexPast].payload;
 
         // Compute the validity of the reports.
         bool recentReportUsable = recentReportTimestamp <= maxValidTimestamp
@@ -372,8 +360,8 @@ contract Oracle is TSOwnable, IOracle {
                 // Note that an average computation of (a + b) / 2 could overflow.
                 // Therefore the computation is distributed:
                 //      (a / 2) + (b / 2) + (((a % 2) + (b % 2)) / 2).
-                uint256 average = (recentReportPayload >> 1)
-                    + (pastReportPayload >> 1)
+                uint average =
+                    (recentReportPayload >> 1) + (pastReportPayload >> 1)
                     + (((recentReportPayload % 2) + (pastReportPayload % 2)) >> 1);
 
                 return (average, true);
